add crud banner

+++ client/src/store/index.js
@@ -11,7 +11,8 @@ export default new Vuex.Store({
     categories: [],
     products: [],
     banner: [],
-    category: ''
+    category: '',
+    bannerStatus: 'All'
   },
   mutations: {
     passingProducts (state, payload) {
@@ -22,6 +23,12 @@ export default new Vuex.Store({
     },
     changeCategory (state, payload) {
       state.category = payload
+    },
+    changeBannerStatus (state, payload) {
+      state.bannerStatus = payload
+    },
+    passingBanner (state, payload) {
+      state.banner = payload
     }
   },
   actions: {
@@ -76,6 +83,25 @@ export default new Vuex.Store({
         })
       })
     },
+    getBanner (context) {
+      axios({
+        method: 'get',
+        url: '/banner',
+        headers: {
+          access_token: localStorage.access_token
+        }
+      }).then(response => {
+        context.commit('passingBanner', response.data)
+      }).catch(err => {
+        console.log(err.response.data.errors)
+        Swal.fire({
+          icon: 'error',
+          title: 'Oops...',
+          text: 'Error while getting banner list',
+          footer: '<a >This might be caused by bad request</a>'
+        })
+      })
+    },
     getCategories (context) {
       axios({
         method: 'get',
@@ -94,8 +120,6 @@ export default new Vuex.Store({
           footer: '<a >This might be caused by bad request</a>'
         })
       })
-    },
-    getBanner (context, payload) {
     },
     addProduct (context, payload) {
       console.log(payload)
@@ -138,6 +162,44 @@ export default new Vuex.Store({
           })
         })
     },
+    addBanner (context, payload) {
+      console.log(payload)
+      axios({
+        method: 'post',
+        url: '/banner',
+        data: {
+          title: payload[0],
+          status: true,
+          image_url: payload[1]
+        },
+        headers: {
+          access_token: localStorage.access_token
+        }
+      })
+        .then((response) => {
+          this.dispatch('getProducts')
+          const answers = JSON.stringify(payload)
+          Swal.fire({
+            title: 'Banner Added!',
+            html: `
+              Your Input:
+              <pre><code>${answers}</code></pre>
+            `,
+            confirmButtonText: 'Nice!'
+          })
+        })
+        .catch((err) => {
+          const errors = err.response.data.errors
+          errors.forEach(element => {
+            Swal.fire({
+              icon: 'error',
+              title: 'Oops...',
+              text: `${element}`,
+              footer: '<a >please follow the rules</a>'
+            })
+          })
+        })
+    },
     delProduct (context, payload) {
       const id = payload
       axios({
@@ -155,6 +217,35 @@ export default new Vuex.Store({
           timer: 1500
         })
         this.dispatch('getProducts')
+      }).catch(err => {
+        const errors = err.response.data.errors
+        errors.forEach(element => {
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: `${element}`,
+            footer: '<a >please follow the rules</a>'
+          })
+        })
+      })
+    },
+    delBanner (context, payload) {
+      const id = payload
+      axios({
+        method: 'delete',
+        url: `/banner/${id}`,
+        headers: {
+          access_token: localStorage.access_token
+        }
+      }).then(response => {
+        Swal.fire({
+          position: 'center',
+          icon: 'success',
+          title: 'Banner Deleted Successfully',
+          showConfirmButton: false,
+          timer: 1500
+        })
+        this.dispatch('getBanner')
       }).catch(err => {
         const errors = err.response.data.errors
         errors.forEach(element => {
@@ -206,6 +297,41 @@ export default new Vuex.Store({
         })
       })
     },
+    updateBanner (context, payload) {
+      const id = payload.id
+      axios({
+        method: 'put',
+        url: `/banner/${id}`,
+        data: {
+          title: payload.name,
+          image_url: payload.image_url,
+          status: payload.status
+        },
+        headers: {
+          access_token: localStorage.access_token
+        }
+      }).then(response => {
+        Swal.fire({
+          position: 'center',
+          icon: 'success',
+          title: 'Success Updating Banner Data',
+          showConfirmButton: false,
+          timer: 1500
+        })
+        // console.log(response);
+        this.dispatch('getBanner')
+      }).catch(err => {
+        const errors = err.response.data.errors
+        errors.forEach(element => {
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: `${element}`,
+            footer: '<a >please follow the rules</a>'
+          })
+        })
+      })
+    },
     patchProduct (context, payload) {
       const { id, categoryId } = payload
       axios({
@@ -238,9 +364,57 @@ export default new Vuex.Store({
         })
       })
     },
+    patchBanner (context, payload) {
+      const { id, status } = payload
+      axios({
+        method: 'patch',
+        url: `/banner/${id}`,
+        data: {
+          status: status
+        },
+        headers: {
+          access_token: localStorage.access_token
+        }
+      }).then(response => {
+        Swal.fire({
+          position: 'center',
+          icon: 'success',
+          title: 'Banner data updated',
+          showConfirmButton: false,
+          timer: 1500
+        })
+        this.dispatch('getBanner')
+      }).catch(err => {
+        const errors = err.response.data.errors
+        errors.forEach(element => {
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: `${element}`,
+            footer: '<a >please follow the rules</a>'
+          })
+        })
+      })
+    },
     logout (context) {
-      localStorage.clear()
-      router.push({ path: 'Login' })
+      Swal.fire({
+        title: 'Are you sure?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Logout'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          localStorage.clear()
+          router.push({ path: 'Login' })
+          Swal.fire(
+            'Logged out!',
+            'You had been log out.',
+            'success'
+          )
+        }
+      })
     },
     uploadImage (context, payload) {
       console.log(payload)
@@ -256,6 +430,9 @@ export default new Vuex.Store({
       }).catch(err => {
         console.log(err)
       })
+    },
+    route (context, payload) {
+      router.push({ path: payload })
     }
   },
   modules: {
@@ -267,6 +444,19 @@ export default new Vuex.Store({
       } else {
         return state.products.filter((value) => {
           return value.categoryId === state.category
+        })
+      }
+    },
+    filterBanner: state => {
+      if (state.bannerStatus === 'All' || state.bannerStatus === '') {
+        return state.banner
+      } else if (state.bannerStatus === 'True') {
+        return state.banner.filter((value) => {
+          return value.status === true
+        })
+      } else if (state.bannerStatus === 'False') {
+        return state.banner.filter((value) => {
+          return value.status === false
         })
       }
     }
